+++ packages/cheqd-blockchain-modules/src/offchain-signatures/internal.js
@@ -3,7 +3,7 @@ import injectParams from '../common/inject-params';
 export default class CheqdOffchainSignaturesInternalModule extends injectParams(
   class {},
 ) {
-  filterParamsMetadata(meta) {
-    return meta.resourceType === 'offchain-signature-params';
-  }
+  static ParamsName = 'OffchainParams';
+
+  static ParamsType = 'offchain-signature-params';
 }