+++ packages/credential-sdk/src/types/accumulator/keys.js
@@ -1,5 +1,5 @@
 import { TypedTuple, TypedNumber } from '../generic';
-import { DockDidOrDidMethodKey } from '../did';
+import { CheqdDLRRef, DockDidOrDidMethodKey } from '../did';
 
 export class DockAccumulatorPublicKeyRef extends TypedTuple {
   static Classes = [DockDidOrDidMethodKey, TypedNumber];
@@ -8,3 +8,7 @@ export class DockAccumulatorPublicKeyRef extends TypedTuple {
 export class DockAccumulatorParamsRef extends TypedTuple {
   static Classes = [DockDidOrDidMethodKey, TypedNumber];
 }
+
+export class CheqdAccumulatorPublicKeyRef extends CheqdDLRRef {}
+
+export class CheqdAccumulatorParamsRef extends CheqdDLRRef {}