+++ packages/credential-sdk/src/types/generic/utils.js
@@ -1,0 +1,8 @@
+import TypedBytes from './typed-bytes';
+import TypedArray from './typed-array';
+
+export class ByteArray extends TypedBytes {}
+
+export class ArrayOfByteArrays extends TypedArray {
+  static Class = TypedBytes;
+}