+++ packages/dock-blockchain-modules/src/accumulator/actions.js
@@ -9,9 +9,10 @@ import {
 import {
   TypedBytes,
   TypedStruct,
-  TypedArray,
   option,
   TypedNumber,
+  ArrayOfByteArrays,
+  ByteArray,
 } from '@docknetwork/credential-sdk/types/generic';
 
 export class AddAccumulator extends TypedStruct {
@@ -22,18 +23,12 @@ export class AddAccumulator extends TypedStruct {
   };
 }
 
-class ByteArray extends TypedBytes {}
-
-export class ListOfByteArrays extends TypedArray {
-  static Class = ByteArray;
-}
-
 export class UpdateAccumulator extends TypedStruct {
   static Classes = {
     id: DockAccumulatorIdIdent,
     newAccumulated: class Accumulated extends TypedBytes {},
-    additions: option(ListOfByteArrays),
-    removals: option(ListOfByteArrays),
+    additions: option(ArrayOfByteArrays),
+    removals: option(ArrayOfByteArrays),
     witnessUpdateInfo: option(ByteArray),
     nonce: TypedNumber,
   };